Add back smallvec.
